Core.cljs: fixed a bug where serve-new-cards would prepend a new card on a certain column instead of appending it.
Problem: discard an ace on a column with at least 3 cards.
Ace gets discarded, card under it gets opened. Then click "Hit me" and serve-new-cards will append cards to all columns but the one where we just removed the ace.
Cause: map-indexed converted the column's :cards attribute to a list (from a vector) and conj would then perpend instead of append the new card.
Solution: replace with the more readable assoc-in, which keeps the vector.

A big thank you to @the-kenny for excellent very late night debugging support!